requirements.txt
asgiref==3.4.1
dj-static==0.0.6
Django==3.2.8
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
static3==0.7.0
typing-extensions==3.10.0.2
whitenoise==5.3.0
